removed ubuntu build from workflow
The primary reason for this to b/c it builds on
ubuntu-latest which does not produce a distributatable
binary (it can only produce a platform specific binary)

To build a manylinux release in the future, it would
be necessary to use the manylinux contianers. There is
some support for this in scikit-build, but for now it's
largely undocumented.